Add files via upload
The project  EdComm is a platform for educators in Bangladesh, enabling them to share and access tailored teaching resources like lesson plans, videos, and presentations. It prioritizes user-friendly design and collaboration, with plans for growth, partnerships, and integration into the education system. EdComm aims to empower teachers and improve resource accessibility in schools nationwide. Developed with HTML, CSS, JavaScript,PHP